Improves ExternalLink's icon usability
These changes make the component's interface align to that of the Button
component, and allow for an icon to be placed at the beginning or end of
the block content.

This does not contain any breaking changes. The component still defaults
to rendering the exit icon at the end of the block content as it did
prior.